backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.30.1
httpx==0.27.0
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==0.6.23
reportlab==4.2.0
aiofiles==23.2.1
pytest==8.2.0
pytest-asyncio==0.23.6
